test(drag-and-drop): correct file naming
